Restructure packaging for redirect
COAND-497
