+++ redirect/src/main/java/com/adyen/checkout/redirect/internal/ui/RedirectDelegate.kt
@@ -6,13 +6,15 @@
  * Created by josephj on 18/8/2022.
  */
 
-package com.adyen.checkout.redirect
+package com.adyen.checkout.redirect.internal.ui
 
+import androidx.annotation.RestrictTo
 import com.adyen.checkout.components.base.ActionDelegate
 import com.adyen.checkout.components.base.DetailsEmittingDelegate
 import com.adyen.checkout.components.base.IntentHandlingDelegate
 import com.adyen.checkout.components.ui.ViewProvidingDelegate
 
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 interface RedirectDelegate :
     ActionDelegate,
     DetailsEmittingDelegate,
